services/chatbot-api/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.30,<1
pydantic>=2.7,<3
implicit>=0.7.2
scikit-learn
numpy
scipy
python-dotenv
oracledb>=2.2,<3




